Refactor to move each tab to their own page
