+++ src/routes/courses/[code]/graphs/[graphid]/+layout.server.ts
@@ -1,0 +1,66 @@
+import prisma from '$lib/server/db/prisma';
+import { GraphValidator } from '$lib/validators/graphValidator';
+import { error } from '@sveltejs/kit';
+import type { LayoutServerLoad } from './$types';
+
+export const load: LayoutServerLoad = async ({ params }) => {
+	if (!params.code || !params.graphid) {
+		error(400, { message: 'Course code and Graph id are required' });
+	}
+
+	const graphId = parseInt(params.graphid);
+	// Check if graphId is not NaN
+	if (isNaN(graphId)) {
+		error(400, { message: 'Graph id must be a number' });
+	}
+
+	try {
+		const course = await prisma.course.findFirst({
+			where: {
+				code: params.code
+			},
+			include: {
+				graphs: {
+					where: {
+						id: graphId
+					},
+					include: {
+						domains: {
+							include: {
+								incommingDomains: true,
+								outgoingDomains: true
+							},
+							orderBy: {
+								order: 'asc'
+							}
+						},
+						subjects: {
+							include: {
+								incommingSubjects: true,
+								outgoingSubjects: true,
+								domain: true
+							}
+						},
+						lectures: true
+					}
+				}
+			}
+		});
+
+		if (!course) error(404, { message: 'Course not found' });
+		if (course.graphs.length === 0) error(404, { message: 'Graph not found' });
+
+		const graphValidator = new GraphValidator(course.graphs[0]);
+
+		const cycles = graphValidator.hasCycle();
+
+		// Happy path
+		return {
+			course: course,
+
+			cycles: cycles
+		};
+	} catch (e: unknown) {
+		error(500, { message: e instanceof Error ? e.message : `${e}` });
+	}
+};
