+++ src/routes/courses/[code]/graphs/[graphid]/domains/+page.server.ts
@@ -1,0 +1,31 @@
+import { DomainActions } from '$lib/server/helpers';
+import {
+	changeDomainRelSchema,
+	deleteDomainSchema,
+	domainRelSchema,
+	domainSchema
+} from '$lib/zod/domainSubjectSchema';
+import type { ServerLoad } from '@sveltejs/kit';
+import { superValidate } from 'sveltekit-superforms';
+import { zod } from 'sveltekit-superforms/adapters';
+
+export const load: ServerLoad = async ({ params }) => {
+	return {
+		newDomainForm: await superValidate(zod(domainSchema)),
+		deleteDomainForm: await superValidate(zod(deleteDomainSchema)),
+		newDomainRelForm: await superValidate(zod(domainRelSchema)),
+		changeDomainRelForm: await superValidate(zod(changeDomainRelSchema))
+	};
+};
+
+// ACTIONS
+export const actions = {
+	'add-domain-to-graph': DomainActions.addDomainToGraph,
+	'change-domain-in-graph': DomainActions.changeDomain,
+	'delete-domain': DomainActions.deleteDomain,
+
+	// Domain relationships
+	'add-domain-rel': DomainActions.addDomainRel,
+	'change-domain-rel': DomainActions.changeDomainRel,
+	'delete-domain-rel': DomainActions.deleteDomainRel
+};
