+++ src/routes/courses/[code]/graphs/[graphid]/domains/DeleteDomain.svelte
