+++ src/routes/courses/[code]/graphs/[graphid]/domains/DomainRelSettings.svelte
