+++ src/routes/courses/[code]/graphs/[graphid]/lectures/+page.svelte
@@ -1,0 +1,1 @@
+<h2>Lectures</h2>
