+++ src/routes/courses/[code]/graphs/[graphid]/subjects/DeleteSubject.svelte
